requirements.txt
altair==5.0.1
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
dnspython==2.3.0
easyocr==1.7.0
executing==1.2.0
Faker==18.10.1
favicon==0.7.0
filelock==3.12.2
fonttools==4.40.0
fuzzywuzzy==0.18.0
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
htbuilder==0.6.1
idna==3.4
imageio==2.31.1
importlib-metadata==6.6.0
ipykernel==6.23.2
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
lazy_loader==0.2
Levenshtein==0.21.1
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==9.1.0
mpmath==1.3.0
mysql-connector-python==8.0.33
nest-asyncio==1.5.6
networkx==3.1
ninja==1.11.1
nltk==3.8.1
numpy==1.24.3
opencv-python-headless==4.7.0.72
packaging==23.1
pandas==2.0.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.3
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
pyap==0.3.1
pyarrow==12.0.0
pyclipper==1.3.0.post4
pydeck==0.8.1b0
Pygments==2.15.1
pymdown-extensions==10.0.1
pymongo==4.3.3
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-bidi==0.4.2
python-dateutil==2.8.2
python-Levenshtein==0.21.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.0
rapidfuzz==3.1.1
regex==2023.6.3
requests==2.31.0
rich==13.4.2
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
shapely==2.0.1
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
SQLAlchemy==2.0.16
st-annotated-text==4.0.0
stack-data==0.6.2
streamlit==1.23.1
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.5
streamlit-embedcode==0.1.2
streamlit-extras==0.2.7
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.3
streamlit-keyup==0.2.0
streamlit_lottie==0.0.5
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.1.0
tifffile==2023.4.12
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.3
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
XlsxWriter==3.1.2
zipp==3.15.0
